adding node attribution to detailed pnr and completing the code restructuring
